Remove .sbt root project identification.
In multi-project builds, .sbt files will live at the root of each
subproject, but sbt has to be run from the root to have access to root
project plugins and settings. Eliminating the not null check for .sbt
allows the sbt run to reach the root of the project without bailing out
too early, which would cause ensime-sbt to fail to load from the root of
a project.